Integrate opa in stack handler
