+++ stackl/application/manager/__init__.py
@@ -2,6 +2,7 @@ from abc import ABC
 
 from datastore.datastore_factory import DataStoreFactory
 
+
 class Manager(ABC):
 
     def __init__(self, manager_factory):
