+++ stackl/application/model/configs/stack_infrastructure_template_model.py
@@ -4,6 +4,7 @@ from pydantic import BaseModel
 
 from model.configs.infrastructure_target_model import InfrastructureTarget
 
+
 class StackInfrastructureTemplate(BaseModel):
     name: str
     description: str = ""
