+++ stackl/application/routers/policies_router.py
@@ -1,15 +1,11 @@
 import logging
-from typing import List
 
-from fastapi import APIRouter, HTTPException, Query
+from fastapi import APIRouter, Query
 
-from globals import document_types
-from utils.general_utils import get_hostname
+from manager.manager_factory import ManagerFactory
 from model.configs.policy_model import Policy
 from model.configs.stack_application_template_model import StackApplicationTemplate
 from model.configs.stack_infrastructure_template_model import StackInfrastructureTemplate
-
-from manager.manager_factory import ManagerFactory
 from opa_broker.opa_broker_factory import OPABrokerFactory
 
 logger = logging.getLogger(__name__)
@@ -26,11 +22,14 @@ queries_for_sit_sats_set = Query(
         "solution_set_per_service",
         "services_targets_resources_match",
         "services_target_resources_no_match"
-          ]
+    ]
 )
 
+
 @router.get('/query')
-def run_query_for_sit_and_sat(sit_name: str = "stack_infrastructure_template_opa_test1", sat_name: str = "stack_application_template_opa_test1", query: str = queries_for_sit_sats_set):
+def run_query_for_sit_and_sat(sit_name: str = "stack_infrastructure_template_opa_test1",
+                              sat_name: str = "stack_application_template_opa_test1",
+                              query: str = queries_for_sit_sats_set):
     sit_doc = document_manager.get_stack_infrastructure_template(sit_name)
     sat_doc = document_manager.get_stack_application_template(sat_name)
     sit_as_opa_data = opa_broker.convert_sit_to_opa_data(sit_doc)
@@ -62,6 +61,7 @@ def get_sat_policies_from_sat(sat_name: str = "stack_application_template_opa_te
     sat_doc = document_manager.get_stack_application_template(sat_name)
     return opa_broker.convert_sat_to_opa_policies(sat_doc)
 
+
 @router.get('/data')
 def get_data(data_path: str = "default"):
     result = opa_broker.get_opa_data(data_path)
@@ -69,7 +69,7 @@ def get_data(data_path: str = "default"):
 
 
 @router.put('/data')
-def put_data(data:dict, data_path: str = "default"):
+def put_data(data: dict, data_path: str = "default"):
     result = opa_broker.load_opa_data(data, data_path)
     return result
 
@@ -82,9 +82,10 @@ def delete_data(data_path: str = "default"):
 
 @router.put('/data/sit')
 def put_sit_data(data: dict, data_path: str = "default"):
-### WIP
+    ### WIP
     result = opa_broker.load_opa_data(data_path)
     return result
+
 
 @router.get('')
 def get_policies():
@@ -103,6 +104,7 @@ def delete_policy(policy_id: str):
     result = opa_broker.delete_opa_policy(policy_id)
     return result
 
+
 @router.put('')
 def put_policy(policy_doc: Policy):
     result = opa_broker.load_opa_policy(policy_doc)
@@ -114,6 +116,7 @@ def put_sit_policies(sit_doc: StackInfrastructureTemplate):
     ### WIP
     result = opa_broker.load_opa_policies_from_sit(sit_doc)
     return result
+
 
 @router.put('/sat')
 def put_sat_policies(sat_doc: StackApplicationTemplate):
